Make it optional to list orig_instr in the JSON file
We want to print orig_inst in the JSON file for verifive sw. But we
want to keep this optional so that it is possible to generate the JSON
file version which can be compared against the version generated by
the master branch. This way it becomes possible to ensure sanity of
the instruction database.